pyproject.toml
[tool.poetry]
name = "bakllava-rlhf"
version = "0.1.0"
description = ""
authors = ["Fraser Mince <frasermince@gmail.com>"]
readme = "README.md"
packages = [{include = "bakllava_rlhf"}]

[tool.poetry.dependencies]
python = "^3.10"
transformers={ version = "4.40.0"}
datasets = "^2.16.1"
trl = "^0.8.1"
torch = "^2.1.2"
peft = "^0.10.0"
bitsandbytes = "^0.42.0"
pillow = "^10.2.0"
nvidia-ml-py3 = "^7.352.0"
accelerate = "^0.29.2"
wandb = "^0.16.6"
flash-attn = { version = "^2.5.7"}
rewardbench = { version = "0.1.1"}

[tool.poetry.optional-dependencies]
flash-attn = { version = "^2.5.7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
